updated project

+++ src/main/java/com/retailmanagement/modules/inventory/model/Inventory.java
@@ -10,6 +10,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -55,10 +56,10 @@ public class Inventory {
     private String shelfNumber;
 
     @Column(precision = 10, scale = 2)
-    private Double averageCost;
+    private BigDecimal averageCost;
 
     @Column(precision = 10, scale = 2)
-    private Double lastPurchasePrice;
+    private BigDecimal lastPurchasePrice;
 
     private LocalDateTime lastStockTakeDate;
